random num type changed to int, but problem of incorrect number remains

+++ code/src/cipher.c
@@ -1,7 +1,7 @@
 #include <stdint.h>
 #include <stdio.h>
 
-void xor_encrypt(uint8_t* plaintext, uint8_t* ciphertext, int message_length, const uint8_t* key, int keysize) {
+void xor_encrypt(int* plaintext, int* ciphertext, int message_length, const int* key, int keysize) {
     int chunk_size = 1, nthreads, thread_id, i, j;
     #pragma omp parallel shared(nthreads, chunk_size, ciphertext) private(thread_id, i, j)
     {
@@ -14,7 +14,7 @@ void xor_encrypt(uint8_t* plaintext, uint8_t* ciphertext, int message_length, co
     }
  
 };
-void xor_decrypt(uint8_t* ciphertext, uint8_t* plaintext, int message_length, const uint8_t* key, int keysize) {
+void xor_decrypt(int* ciphertext, int* plaintext, int message_length, const int* key, int keysize) {
     xor_encrypt(ciphertext, plaintext, message_length, key, keysize);
 };
